+++ packages/assets-ergot/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @happyvertical/smrt-assets-ergot
 
+## 0.24.14
+
+### Patch Changes
+
+- @happyvertical/smrt-assets@0.24.14
+
+## 0.24.13
+
+### Patch Changes
+
+- @happyvertical/smrt-assets@0.24.13
+
 ## 0.24.12
 
 ### Patch Changes