+++ packages/assets/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @happyvertical/smrt-assets
 
+## 0.24.14
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.24.14
+  - @happyvertical/smrt-tags@0.24.14
+  - @happyvertical/smrt-tenancy@0.24.14
+  - @happyvertical/smrt-svelte@0.24.14
+  - @happyvertical/smrt-types@0.24.14
+
+## 0.24.13
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.24.13
+  - @happyvertical/smrt-tags@0.24.13
+  - @happyvertical/smrt-tenancy@0.24.13
+  - @happyvertical/smrt-svelte@0.24.13
+  - @happyvertical/smrt-types@0.24.13
+
 ## 0.24.12
 
 ### Patch Changes