+++ packages/content/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @happyvertical/smrt-content
 
+## 0.24.14
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.24.14
+  - @happyvertical/smrt-assets@0.24.14
+  - @happyvertical/smrt-chat@0.24.14
+  - @happyvertical/smrt-facts@0.24.14
+  - @happyvertical/smrt-images@0.24.14
+  - @happyvertical/smrt-messages@0.24.14
+  - @happyvertical/smrt-profiles@0.24.14
+  - @happyvertical/smrt-prompts@0.24.14
+  - @happyvertical/smrt-tenancy@0.24.14
+  - @happyvertical/smrt-types@0.24.14
+
+## 0.24.13
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.24.13
+  - @happyvertical/smrt-assets@0.24.13
+  - @happyvertical/smrt-chat@0.24.13
+  - @happyvertical/smrt-facts@0.24.13
+  - @happyvertical/smrt-images@0.24.13
+  - @happyvertical/smrt-messages@0.24.13
+  - @happyvertical/smrt-profiles@0.24.13
+  - @happyvertical/smrt-prompts@0.24.13
+  - @happyvertical/smrt-tenancy@0.24.13
+  - @happyvertical/smrt-types@0.24.13
+
 ## 0.24.12
 
 ### Patch Changes