+++ packages/features/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @happyvertical/smrt-features
 
+## 0.24.14
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.24.14
+  - @happyvertical/smrt-users@0.24.14
+
+## 0.24.13
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.24.13
+  - @happyvertical/smrt-users@0.24.13
+
 ## 0.24.12
 
 ### Patch Changes