+++ packages/smrt-svelte/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @happyvertical/smrt-svelte
 
+## 0.24.14
+
+### Patch Changes
+
+- @happyvertical/smrt-agents@0.24.14
+- @happyvertical/smrt-jobs@0.24.14
+- @happyvertical/smrt-profiles@0.24.14
+- @happyvertical/smrt-users@0.24.14
+- @happyvertical/smrt-types@0.24.14
+
+## 0.24.13
+
+### Patch Changes
+
+- @happyvertical/smrt-agents@0.24.13
+- @happyvertical/smrt-jobs@0.24.13
+- @happyvertical/smrt-profiles@0.24.13
+- @happyvertical/smrt-users@0.24.13
+- @happyvertical/smrt-types@0.24.13
+
 ## 0.24.12
 
 ### Patch Changes