+++ packages/tags/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @happyvertical/smrt-tags
 
+## 0.24.14
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.24.14
+  - @happyvertical/smrt-tenancy@0.24.14
+
+## 0.24.13
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.24.13
+  - @happyvertical/smrt-tenancy@0.24.13
+
 ## 0.24.12
 
 ### Patch Changes